update blood aegis and bonfire formations
