+++ source/gear/_gearDictionary.js
@@ -15,12 +15,8 @@ for (const file of [
 	"_greed.js",
 	"arcane-sledge.js",
 	"battle-standard.js",
-	"bloodaegis-base.js",
-	"bloodaegis-toxic.js",
-	"bloodaegis-urgent.js",
-	"bonfireformation-base.js",
-	"bonfireformation-charging.js",
-	"bonfireformation-hastening.js",
+	"blood-aegis.js",
+	"bonfire-formationjs",
 	"bootsofcomfort-accurate.js",
 	"bootsofcomfort-base.js",
 	"bootsofcomfort-hearty.js",
